modificacion del tipo de variable en ejercicio 6

+++ Ejercicio6.java
@@ -5,11 +5,11 @@ public class Ejercicio6{
     System.out.println("Ingrese n numero reales y cuando coloque un 0 mostrar la suma por pantalla");
     // para leer por teclado
     Scanner teclado=new Scanner(System.in);
-    double valorIngresado=teclado.nextDouble();
+    float valorIngresado=teclado.nextFloat();
     // hasta aca
-    double resultado=valorIngresado; // ya empiezo almacenar el valor de teclado
+    float resultado=valorIngresado; // ya empiezo almacenar el valor de teclado
     for (;valorIngresado!=0;){
-      valorIngresado=teclado.nextDouble();
+      valorIngresado=teclado.nextFloat();
       resultado=resultado+valorIngresado; // para poder salir del for la variable valorIngresado almacena lo que carga en teclado   
     }
     System.out.println("La suma de de los numeros es " + resultado);
